requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==9.0.0
pyaml==21.10.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
